build: Reject PR if requirements files are inconsistent
We have a few situations where requirements files can become inconsistent
or cause unnecessary review churn in later `make upgrade` runs. This will
produce a failing check for any PR that does not produce a clean run of
`make compile-requirements` on Linux.

This also changes `make compile-requirements` to not try upgrading the
pip version file. This is necessary for diff stability, and it wasn't
installing the upgraded version anyhow.

Closes https://github.com/openedx/edx-platform/issues/31372